Update python (#18)
* Update Matplotlib to 3.9.x

This changes the number of ticks of some plots as described in the changelog: https://matplotlib.org/stable/api/prev_api_changes/api_changes_3.8.0.html#placing-of-maximum-and-minimum-minor-ticks

* Updated Python libraries and regenerated the plots

This slightly changes the plots as the libraries now use a different version of the floating point library. These subtle changes do not affect the results of the thesis and are negligible.