Bump TexLive to 2023
This fixes https://github.com/T-F-S/tcolorbox/issues/123 and removes a space, thereby changing the height of all colorboxes like the 'Requirements' boxes used throughout the preparation chapter.

There are also changes to tikz-base which is used for drawing some graphs. This fixes the alignment of the tick labels. They are now correctly centered below the tick.

The microtyping package seems to compress the text a little more, so that long passages tend to take less space.